+++ app/src/main/java/com/example/mf4z/fulfulde/ColorsActivity.java
@@ -24,7 +24,6 @@ public class ColorsActivity extends AppCompatActivity {
                 .replace(R.id.container, new ColorsFragment())
                 .commit();
 
-
     }
